# Subquery Spin-Off: Optimize "enable all" regexp