This feature has been enabled in all of our cells with good results. I use `.*` as an enable-all flag, but I realized that it runs a regexp thousands of times per second for no reason

A boolean would probably be better but changing the flag would be too much of a hassle at this point

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/frontend/querymiddleware
cpu: Apple M3 Pro
                                                                         │ benchmark_results1.txt │       benchmark_results2.txt        │
                                                                         │         sec/op         │    sec/op     vs base               │
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_simple_rate-11                      132.9n ± 2%   136.9n ±  3%   +3.01% (p=0.013 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_simple_sum-11                       132.9n ± 5%   133.2n ±  2%        ~ (p=0.589 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_complex_query-11                    136.9n ± 4%   134.8n ±  1%        ~ (p=0.394 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/match_all_simple_rate-11                        775.5n ± 4%   137.2n ±  3%  -82.30% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/match_all_simple_sum-11                         899.2n ± 4%   135.2n ±  1%  -84.96% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/match_all_complex_query-11                     1787.5n ± 4%   137.9n ±  5%  -92.29% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_simple_rate-11                   789.5n ± 1%   785.2n ±  3%        ~ (p=0.511 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_simple_sum-11                    901.6n ± 2%   917.9n ±  3%        ~ (p=0.132 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_complex_query-11                 1.807µ ± 3%   1.796µ ± 14%        ~ (p=0.937 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_simple_rate-11                2.700µ ± 1%   2.748µ ± 37%        ~ (p=0.167 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_simple_sum-11                 3.118µ ± 0%   3.130µ ±  2%        ~ (p=0.076 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_complex_query-11              5.976µ ± 1%   6.048µ ±  3%        ~ (p=0.093 n=6)
geomean                                                                               872.7n        523.5n        -40.01%

                                                                         │ benchmark_results1.txt │        benchmark_results2.txt         │
                                                                         │          B/op          │     B/op      vs base                 │
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_simple_rate-11                       16.00 ± 0%     16.00 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_simple_sum-11                        16.00 ± 0%     16.00 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_complex_query-11                     16.00 ± 0%     16.00 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/match_all_simple_rate-11                        256.00 ± 0%     16.00 ± 0%  -93.75% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/match_all_simple_sum-11                         336.00 ± 0%     16.00 ± 0%  -95.24% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/match_all_complex_query-11                      816.00 ± 0%     16.00 ± 0%  -98.04% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_simple_rate-11                    256.0 ± 0%     256.0 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_simple_sum-11                     336.0 ± 0%     336.0 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_complex_query-11                  816.0 ± 0%     816.0 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_simple_rate-11                 744.0 ± 0%     744.0 ± 1%        ~ (p=0.485 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_simple_sum-11                  987.5 ± 0%     986.5 ± 0%        ~ (p=0.426 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_complex_query-11             2.385Ki ± 0%   2.384Ki ± 0%        ~ (p=1.000 n=6)
geomean                                                                                239.8          106.4       -55.63%
¹ all samples are equal

                                                                         │ benchmark_results1.txt │       benchmark_results2.txt        │
                                                                         │       allocs/op        │ allocs/op   vs base                 │
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_simple_rate-11                       1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_simple_sum-11                        1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/no_patterns_complex_query-11                     1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/match_all_simple_rate-11                        13.000 ± 0%   1.000 ± 0%  -92.31% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/match_all_simple_sum-11                         16.000 ± 0%   1.000 ± 0%  -93.75% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/match_all_complex_query-11                      36.000 ± 0%   1.000 ± 0%  -97.22% (p=0.002 n=6)
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_simple_rate-11                    13.00 ± 0%   13.00 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_simple_sum-11                     16.00 ± 0%   16.00 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/simple_pattern_complex_query-11                  36.00 ± 0%   36.00 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_simple_rate-11                 37.00 ± 0%   37.00 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_simple_sum-11                  46.00 ± 0%   46.00 ± 0%        ~ (p=1.000 n=6) ¹
SpinOffSubqueriesMiddleware_IsEnabled/multiple_patterns_complex_query-11               106.0 ± 0%   106.0 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                                                12.13        5.766       -52.45%
¹ all samples are equal
```

<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [x] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.