@56quarters 2025-03-31
> A boolean would probably be better but changing the flag would be too much of a hassle at this point

If it's still experimental my preference would be to deal with the hassle now before we mark the feature as stable. If we're using the wildcard regex everywhere it should be easy enough to make the option format conditional on weekly release, no?