@julienduchesne 2025-04-08
> Approved but I think we should consider an on/off flag if that's what we end up doing most of the time.

Superseded by https://github.com/grafana/mimir/pull/11153. Changed the flag to a boolean there. I didn't want to do a 2-release change, which is required when we change per-tenant overrides; but we don't use the per-tenant override, so I can set the new flag at the same time as the new release, not much of a hassle in the end.